% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_crispy.R
\name{aggregate_crispy_facts}
\alias{aggregate_crispy_facts}
\title{Title}
\usage{
aggregate_crispy_facts(multi_crispy, group_cols)
}
\arguments{
\item{multi_crispy}{multi_crispy}

\item{group_cols}{group_cols}
}
\description{
TODO FIND CLOSEST COMPANY IF group_cols=NULL
}
