% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_portfolio.R
\name{aggregate_portfolio_facts}
\alias{aggregate_portfolio_facts}
\title{Title}
\usage{
aggregate_portfolio_facts(portfolio_data, group_cols)
}
\arguments{
\item{portfolio_data}{portfolio_data}

\item{group_cols}{group_cols}
}
\description{
Title
}
