% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_trajectories.R
\name{aggregate_trajectories_facts}
\alias{aggregate_trajectories_facts}
\title{Aggregate numerical trajectories columns}
\usage{
aggregate_trajectories_facts(multi_trajectories, group_cols)
}
\arguments{
\item{multi_trajectories}{dataframe of trajectories from 1 or multiple trisk truns}

\item{group_cols}{group_cols}
}
\description{
Aggregate numerical trajectories columns
}
