% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_trisk_runs.R
\name{collect_and_aggregate_crispys}
\alias{collect_and_aggregate_crispys}
\title{Crispy collect}
\usage{
collect_and_aggregate_crispys(all_runs)
}
\arguments{
\item{all_runs}{}
}
\description{
Crispy collect
}
