% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_analysis.R
\name{compute_analysis_metrics}
\alias{compute_analysis_metrics}
\title{Compute Analysis Metrics}
\usage{
compute_analysis_metrics(analysis_data)
}
\arguments{
\item{analysis_data}{analysis_data}
}
\description{
Function computing financial metrics to use for analysis
}
