% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_trajectories.R
\name{convert_trajectories_as_percentages}
\alias{convert_trajectories_as_percentages}
\title{Convert trajectories from absolute values to percentage of the max}
\usage{
convert_trajectories_as_percentages(multi_trajectories, group_cols)
}
\arguments{
\item{multi_trajectories}{dataframe of trajectories from 1 or multiple trisk truns}

\item{group_cols}{group_cols}
}
\description{
Convert trajectories from absolute values to percentage of the max
}
