% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_analysis.R
\name{create_analysis_data}
\alias{create_analysis_data}
\title{Title}
\usage{
create_analysis_data(
  portfolio_data,
  multi_crispy_data,
  portfolio_crispy_merge_cols
)
}
\arguments{
\item{portfolio_data}{portfolio_data}

\item{multi_crispy_data}{multi_crispy_data}

\item{portfolio_crispy_merge_cols}{portfolio_crispy_merge_cols}
}
\description{
Title
}
