% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_trisk_runs.R
\name{download_experiment_outputs}
\alias{download_experiment_outputs}
\title{Artifacts collect}
\usage{
download_experiment_outputs(all_runs, output_dir, unzip = F)
}
\arguments{
\item{unzip}{}
}
\description{
Artifacts collect
}
