% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispy_pd_term_plot.R
\name{draw_pd_term_plot}
\alias{draw_pd_term_plot}
\title{Generate Plot of PD Values by Term and Scenario}
\usage{
draw_pd_term_plot(data_pd_term, facet_var)
}
\arguments{
\item{data_pd_term}{Prepared data for plotting, containing term-wise PD values and scenarios.}

\item{facet_var}{Variable used to segment data into different business units for faceted visualization.}
}
\value{
A ggplot object depicting the comparative analysis of PD values by term and scenario, crucial for evaluating financial risk and strategic planning.
}
\description{
Creates a bar plot to visualize the variation of PD values across different terms, under baseline and shock scenarios, segmented by business unit. Utilizes a color gradient to distinguish between PD value magnitudes, facilitating a clear and intuitive comparison of risk profiles.
}
