% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispy_trisk_lineplot.R
\name{draw_trisk_line_plot}
\alias{draw_trisk_line_plot}
\title{Draw Line Plot for Risk Trajectories}
\usage{
draw_trisk_line_plot(data_trisk_line_plot, x_var, facet_var, linecolor)
}
\arguments{
\item{data_trisk_line_plot}{Prepared data for plotting, with production percentages and scenarios.}

\item{x_var}{Time variable for the x-axis.}

\item{facet_var}{Variable for faceting plots by business units.}

\item{linecolor}{Variable for coloring lines by sector.}
}
\value{
A ggplot object illustrating risk trajectories over time, aiding in the analysis of production risk and scenario planning.
}
\description{
Creates a line plot to depict the production risk trajectories as a percentage of the maximum value, offering a visual comparison across different scenarios within business units and sectors. Uses color and linetype variations to differentiate between sectors and scenarios.
}
