% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_trisk_runs.R
\name{get_successful_experiment_runs}
\alias{get_successful_experiment_runs}
\title{Title}
\usage{
get_successful_experiment_runs(experiment_name)
}
\arguments{
\item{experiment_name}{}
}
\description{
Title
}
