% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_data_load.R
\name{load_input_plots_data_from_files}
\alias{load_input_plots_data_from_files}
\title{Data load function to generate plots}
\usage{
load_input_plots_data_from_files(
  crispy_outputs_dir,
  portfolio_data_path = NULL,
  granularity = c("company_id", "company_name", "ald_sector", "ald_business_unit"),
  maturity_month_term_bridge_fp = here::here("data", "maturity_month_term_bridge.csv")
)
}
\arguments{
\item{crispy_outputs_dir}{crispy_outputs_dir}

\item{portfolio_data_path}{portfolio_data_path}

\item{granularity}{granularity}

\item{maturity_month_term_bridge_fp}{see in data-raw in the source package}

\item{trisk_start_year}{(default) sets to the earliest year of multi_cripy_data}
}
\description{
The dataframe in output of this function should always be
the one used as input for the plots preprocessing functions
}
