% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_data_load.R
\name{main_data_load_trajectories_data}
\alias{main_data_load_trajectories_data}
\title{Title}
\usage{
main_data_load_trajectories_data(company_trajectories_data, group_cols)
}
\arguments{
\item{company_trajectories_data}{company_trajectories_data}
}
\description{
Title
}
