% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_main.R
\name{main_data_load_trajectories_data_from_file}
\alias{main_data_load_trajectories_data_from_file}
\title{Title}
\usage{
main_data_load_trajectories_data_from_file(
  crispy_outputs_dir,
  granularity,
  param_cols = c("run_id", "year")
)
}
\arguments{
\item{crispy_outputs_dir}{crispy_outputs_dir}

\item{granularity}{granularity}

\item{param_cols}{param_cols}
}
\description{
Title
}
