% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_data_load.R
\name{main_load_analysis_data}
\alias{main_load_analysis_data}
\title{Title}
\usage{
main_load_analysis_data(
  multi_crispy_data,
  portfolio_data = NULL,
  portfolio_crispy_merge_cols = NULL
)
}
\arguments{
\item{portfolio_crispy_merge_cols}{}
}
\description{
The dataframe in output of this function should always be
the one used as input for the plots preprocessing functions
}
