% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_main.R
\name{main_load_portfolio_data}
\alias{main_load_portfolio_data}
\title{Title}
\usage{
main_load_portfolio_data(
  portfolio_data,
  granularity,
  param_cols = c("portfolio_id", "term", "asset_type"),
  trisk_start_year
)
}
\arguments{
\item{portfolio_data}{portfolio_data}

\item{granularity}{granularity}

\item{param_cols}{param_cols}

\item{trisk_start_year}{trisk_start_year}
}
\description{
Title
}
