% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cdi_plots.R
\name{make_density_plots}
\alias{make_density_plots}
\title{Title}
\usage{
make_density_plots(
  data_cdi_pd_plot,
  numeric_values,
  density_var,
  group_variable = "portfolio.ald_sector"
)
}
\arguments{
\item{density_var}{}
}
\description{
Title
}
