% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cdi_plots.R
\name{make_discount_rate_plot}
\alias{make_discount_rate_plot}
\title{Title}
\usage{
make_discount_rate_plot(plot_data, scenario_name_for_title, y_var)
}
\arguments{
\item{y_var}{}
}
\description{
Title
}
