% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_plots.R
\name{make_expected_loss_plot}
\alias{make_expected_loss_plot}
\title{Title}
\usage{
make_expected_loss_plot(
  Scenarios_Sy,
  loans_data,
  scenario_weights,
  groups_var,
  color_gradient
)
}
\arguments{
\item{color_gradient}{}
}
\description{
Title
}
