% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cdi_plots.R
\name{make_mean_pd_diff_plot}
\alias{make_mean_pd_diff_plot}
\title{Title}
\usage{
make_mean_pd_diff_plot(data_cdi_pd_plot, scenario_name_for_title)
}
\arguments{
\item{scenario_name_for_title}{}
}
\description{
Title
}
