% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_portfolio.R
\name{map_portfolio_maturity_to_term}
\alias{map_portfolio_maturity_to_term}
\title{Title}
\usage{
map_portfolio_maturity_to_term(
  portfolio_data,
  maturity_month_term_bridge,
  trisk_start_year = 2022
)
}
\arguments{
\item{start_year}{}
}
\description{
Title
}
