% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispy_exposure_change_plot.R
\name{pipeline_crispy_exposure_change_plot}
\alias{pipeline_crispy_exposure_change_plot}
\title{Visualize Exposure and Value Loss Changes}
\usage{
pipeline_crispy_exposure_change_plot(
  analysis_data,
  x_var = "ald_sector",
  y_exposure_var = "exposure_value_usd",
  y_value_loss_var = "crispy_value_loss",
  facet_var = NULL
)
}
\arguments{
\item{analysis_data}{Dataframe with sector/category-wise financial data.}

\item{x_var}{Variable on the x-axis, typically sector or category.}

\item{y_exposure_var}{Variable for exposure values to be visualized.}

\item{y_value_loss_var}{Variable for crispy value loss to be overlayed.}

\item{facet_var}{Optional; faceting variable to segment data further.}
}
\value{
A ggplot object that shows changes in exposure values and value losses, aiding in risk evaluation and management.
}
\description{
Creates a visualization of exposure value changes alongside crispy value losses for different sectors or categories. It allows for an adjustable focus through faceting and customization of the exposure and loss variables. This plot is vital for stakeholders to assess the impact of various factors on sectoral financial stability and risk exposure.
}
