% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispy_npv_change_plot.R
\name{pipeline_crispy_npv_change_plot}
\alias{pipeline_crispy_npv_change_plot}
\title{Visualize Sector-wise NPV Percentage Changes}
\usage{
pipeline_crispy_npv_change_plot(
  analysis_data,
  x_var = "ald_sector",
  y_var = "crispy_perc_value_change"
)
}
\arguments{
\item{analysis_data}{Dataframe containing sector-wise NPV changes and other financial metrics.}

\item{x_var}{Sector variable for categorization, defaulting to "ald_sector".}

\item{y_var}{NPV change variable, defaulting to "crispy_perc_value_change".}
}
\value{
A ggplot object illustrating the percentage change in NPV across sectors, essential for financial and strategic planning.
}
\description{
Generates a plot to visualize net present value (NPV) percentage changes across sectors, facilitating quick insights into sectors' performance under varying conditions. This function streamlines the data preparation and plotting process, focusing on key financial metrics for strategic analysis.
}
