% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispy_pd_term_plot.R
\name{pipeline_crispy_pd_term_plot}
\alias{pipeline_crispy_pd_term_plot}
\title{Visualize PD Values by Term and Business Unit}
\usage{
pipeline_crispy_pd_term_plot(crispy_data_agg, facet_var = "ald_sector")
}
\arguments{
\item{crispy_data_agg}{Aggregated dataset containing PD values for various terms and scenarios across different sectors or business units.}

\item{facet_var}{The variable by which to facet the plot, typically representing different sectors or business units, defaulting to "ald_sector".}
}
\value{
A ggplot object showing PD values by term, differentiated by scenario, across the specified business units, aiding in strategic risk management.
}
\description{
Orchestrates the creation of a plot visualizing the probability of default (PD) values across different terms and sectors or business units. This function facilitates an understanding of how PD values vary by term under baseline and shock scenarios, enabling stakeholders to assess risk exposure.
}
