% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispy_trisk_lineplot.R
\name{pipeline_crispy_trisk_line_plot}
\alias{pipeline_crispy_trisk_line_plot}
\title{Visualize Trajectory Risks Over Time by Business Unit and Sector}
\usage{
pipeline_crispy_trisk_line_plot(
  trajectories_data,
  x_var = "year",
  facet_var = "ald_business_unit",
  linecolor = "ald_sector"
)
}
\arguments{
\item{trajectories_data}{Dataframe containing yearly data on risk trajectories across different business units and sectors.}

\item{x_var}{The time variable, defaulting to "year".}

\item{facet_var}{The variable for faceting the plot by business units, defaulting to "ald_business_unit".}

\item{linecolor}{Variable determining line colors by sector, defaulting to "ald_sector".}
}
\value{
A ggplot object displaying the trend of risk trajectories over time, providing insights into risk management and strategic planning.
}
\description{
This function generates a line plot visualizing the trajectory of risks over time, segmented by business units and differentiated by sectors. It allows stakeholders to assess trends and shifts in risk levels across different scenarios, facilitating informed decision-making.
}
