% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_el_sector.R
\name{plot_el_coloured}
\alias{plot_el_coloured}
\title{Title}
\usage{
plot_el_coloured(data, y_val_name, is_percentage = FALSE)
}
\arguments{
\item{data}{data}

\item{y_val_name}{y_val_name}

\item{is_percentage}{is_percentage}
}
\description{
Title
}
