% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_val_change.R
\name{plot_value_change}
\alias{plot_value_change}
\title{Title}
\usage{
plot_value_change(data_plt, is_percentage)
}
\arguments{
\item{data_plt}{data_plt}

\item{is_percentage}{is_percentage}
}
\description{
Title
}
