% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_val_change_sector.R
\name{plot_value_change_coloured}
\alias{plot_value_change_coloured}
\title{Title}
\usage{
plot_value_change_coloured(data, is_percentage)
}
\arguments{
\item{data}{The data frame containing the value change data.}

\item{is_percentage}{A logical value indicating whether the value change is in percentage or absolute.}
}
\value{
None
}
\description{
Title
}
