% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_cdi_plots.R
\name{prepare_for_cdi_el_plots}
\alias{prepare_for_cdi_el_plots}
\title{Title}
\usage{
prepare_for_cdi_el_plots(
  analysis_data,
  group_variables_vec = group_variables_vec,
  metrics_el = metrics_el
)
}
\arguments{
\item{analysis_data}{analysis_data}

\item{group_variables_vec}{group_variables_vec}

\item{metrics_el}{metrics_el}
}
\description{
Title
}
