% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_cdi_plots.R
\name{prepare_for_cdi_npv_plots}
\alias{prepare_for_cdi_npv_plots}
\title{Title}
\usage{
prepare_for_cdi_npv_plots(analysis_data, group_variables_vec, metrics_npv)
}
\arguments{
\item{analysis_data}{analysis_data}

\item{group_variables_vec}{group_variables_vec}

\item{metrics_npv}{metrics_npv}
}
\description{
Title
}
