% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_cdi_plots.R
\name{prepare_for_cdi_pd_plots}
\alias{prepare_for_cdi_pd_plots}
\title{Title}
\usage{
prepare_for_cdi_pd_plots(
  analysis_data,
  group_variables_vec,
  weight_variable_char = NULL,
  metrics_pd
)
}
\arguments{
\item{metrics_pd}{}
}
\description{
Title
}
