% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispy_npv_change_plot.R
\name{prepare_for_crispy_npv_change_plot}
\alias{prepare_for_crispy_npv_change_plot}
\title{Prepare Data for NPV Change Visualization}
\usage{
prepare_for_crispy_npv_change_plot(analysis_data, x_var, y_var)
}
\arguments{
\item{analysis_data}{Dataset including sector-wise financial metrics and NPV changes.}

\item{x_var}{Sector categorization variable.}

\item{y_var}{NPV change percentage variable.}
}
\value{
A dataframe with unified sector categorization, ready for NPV change visualization.
}
\description{
Transforms the analysis dataset for visualization, focusing on NPV percentage changes. It ensures that the data is in the correct format for the plotting function, enhancing clarity and focus in the visual representation of NPV changes.
}
