% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_el_plots.R
\name{prepare_for_el_plots}
\alias{prepare_for_el_plots}
\title{Title}
\usage{
prepare_for_el_plots(
  analysis_data_single_run,
  group_variable_char,
  value_to_plot_char
)
}
\arguments{
\item{value_to_plot_char}{}
}
\description{
Title
}
