% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispy_exposure_change_plot.R
\name{prepare_for_exposure_change_plot}
\alias{prepare_for_exposure_change_plot}
\title{Preprocess Data for Exposure and Value Loss Visualization}
\usage{
prepare_for_exposure_change_plot(
  analysis_data,
  x_var,
  y_exposure_var,
  y_value_loss_var
)
}
\arguments{
\item{analysis_data}{Dataset including financial metrics for exposure and value loss.}

\item{x_var}{Category or sector variable.}

\item{y_exposure_var}{Exposure value metric.}

\item{y_value_loss_var}{Crispy value loss metric.}
}
\value{
A simplified dataframe focusing on selected variables for visualization.
}
\description{
Transforms given dataset for visualizing changes in exposure and crispy value losses, selecting only the relevant variables. This step ensures that the visualization is focused and clear, aiding in the analysis of financial risk and impact across sectors or categories.
}
