% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_exposure_plot.R
\name{prepare_for_exposure_plot}
\alias{prepare_for_exposure_plot}
\title{Title}
\usage{
prepare_for_exposure_plot(
  analysis_data_single_run,
  group_variable_char,
  value_to_plot_char
)
}
\arguments{
\item{group_variable_char}{group_variable_char}

\item{value_to_plot_char}{value_to_plot_char}

\item{analysis_data}{analysis_data}
}
\description{
the data in output of this function is NOT aggregated on the group variable
}
