% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_exposure_plot.R
\name{prepare_for_exposure_plots}
\alias{prepare_for_exposure_plots}
\title{Title}
\usage{
prepare_for_exposure_plots(analysis_data, group_variable_vec, agg_fn)
}
\arguments{
\item{value_to_plot_char}{}
}
\description{
the data in output of this function is NOT aggregated on the group variable
}
