% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_pd_annual_plot.R
\name{prepare_for_pd_annual_plot}
\alias{prepare_for_pd_annual_plot}
\title{Title}
\usage{
prepare_for_pd_annual_plot(data, group_variable_charvec, value_to_plot_char)
}
\arguments{
\item{value_to_plot_char}{}
}
\description{
Title
}
