% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crispy_pd_term_plot.R
\name{prepare_for_pd_term_plot}
\alias{prepare_for_pd_term_plot}
\title{Prepare Data for PD Term Visualization}
\usage{
prepare_for_pd_term_plot(crispy_data_agg, facet_var)
}
\arguments{
\item{crispy_data_agg}{Dataset containing aggregated PD values across terms and scenarios.}

\item{facet_var}{Faceting variable representing sectors or business units for detailed comparative analysis.}
}
\value{
A dataframe in long format, ready for plotting PD values by term and scenario.
}
\description{
Prepares the aggregated crispy data for visualization, focusing on PD values by term. It transforms the dataset to a long format suitable for plotting, allowing for a comparative view of PD values under different scenarios.
}
