% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_performers_plot.R
\name{prepare_for_performers_plot}
\alias{prepare_for_performers_plot}
\title{Title}
\usage{
prepare_for_performers_plot(
  analysis_data,
  group_variable_vec,
  performer_variable_char,
  performance_metric,
  n_performers = 10
)
}
\arguments{
\item{performer_variable_char}{ex: "company_id", "ald_business_unit"}

\item{group_variable_char}{ex: "ald_sector"}

\item{value_to_plot_char}{ex:}
}
\description{
Title
}
