% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_value_change_plots.R
\name{prepare_for_value_change_plot}
\alias{prepare_for_value_change_plot}
\title{Title}
\usage{
prepare_for_value_change_plot(
  analysis_data_single_run,
  group_variable_charvec,
  value_to_plot_char
)
}
\arguments{
\item{value_to_plot_char}{"baseline", "shock", or "portfolio"}

\item{analysis_data}{analysis_data}
}
\description{
Title
}
