% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_el_change_sector.R
\name{qplot_el_change_sector_perc}
\alias{qplot_el_change_sector_perc}
\title{Title}
\usage{
qplot_el_change_sector_perc(data_change)
}
\arguments{
\item{data_change}{data_change}
}
\description{
Title
}
