% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_histogram_density.R
\name{qplot_histogram_density}
\alias{qplot_histogram_density}
\title{Title}
\usage{
qplot_histogram_density(data_exposure_plot, plot_variable)
}
\arguments{
\item{plot_variable}{}
}
\description{
Title
}
