% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_pd_sector.R
\name{qplot_pd_sector}
\alias{qplot_pd_sector}
\title{Title}
\usage{
qplot_pd_sector(data_pd, common_y_axis = FALSE, annual_pds = FALSE)
}
\arguments{
\item{data_pd}{data_pd}

\item{common_y_axis}{common_y_axis}

\item{annual_pds}{annual_pds}
}
\description{
Title
}
