% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_pie_summaries.R
\name{qplot_pie_summaries}
\alias{qplot_pie_summaries}
\title{Title}
\usage{
qplot_pie_summaries(data_summary_plot, agg_sum_name_chr)
}
\arguments{
\item{data_summary_plot}{data_summary_plot}

\item{agg_sum_name_chr}{"group_sum" or "group_n"}
}
\description{
Title
}
