% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_val_change_sector.R
\name{qplot_val_change_facet}
\alias{qplot_val_change_facet}
\title{Title}
\usage{
qplot_val_change_facet(data_val_change, facet_chr)
}
\arguments{
\item{data_val_change}{The data frame containing the value change data.}

\item{facet_chr}{The column name to be used for faceting the plots.}
}
\description{
Title
}
