% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_val_change_technology.R
\name{qplot_val_change_technology_perc}
\alias{qplot_val_change_technology_perc}
\title{Title}
\usage{
qplot_val_change_technology_perc(data)
}
\arguments{
\item{data}{}
}
\description{
Title
}
