% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_crispy.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Title}
\usage{
remove_outliers(data)
}
\arguments{
\item{df}{df}

\item{column}{column name}

\item{max_zscore}{max_zscore}
}
\description{
Function to remove outliers based on z-score
}
