% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_crispy.R
\name{remove_outliers_per_group}
\alias{remove_outliers_per_group}
\title{Title}
\usage{
remove_outliers_per_group(multi_crispy, group_cols)
}
\arguments{
\item{multi_crispy}{multi_crispy}

\item{group_cols}{group_cols}
}
\description{
Title
}
