% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_trisk_runs.R
\name{setup_r_mlflow}
\alias{setup_r_mlflow}
\title{In a terminal opened in the project folder,
execute this command to start a local mlflow server :
mlflow server --backend-store-uri ~/mlflow_outputs/mlruns --default-artifact-root ~/mlflow_outputs/mlartifacts --serve-artifacts --host 127.0.0.1 --port 5000}
\usage{
setup_r_mlflow(
  mlflow_python_bin,
  mlflow_bin,
  mlflow_uri = "http://localhost:5000"
)
}
\arguments{
\item{mlflow_uri}{}
}
\description{
In a terminal opened in the project folder,
execute this command to start a local mlflow server :
mlflow server --backend-store-uri ~/mlflow_outputs/mlruns --default-artifact-root ~/mlflow_outputs/mlartifacts --serve-artifacts --host 127.0.0.1 --port 5000
}
