% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_crispy.R
\name{aggregate_crispy_facts}
\alias{aggregate_crispy_facts}
\title{Title}
\usage{
aggregate_crispy_facts(
  multi_crispy,
  group_cols,
  param_cols = c("run_id", "roll_up_type", "baseline_scenario", "shock_scenario",
    "risk_free_rate", "discount_rate", "dividend_rate", "growth_rate", "shock_year")
)
}
\arguments{
\item{group_cols}{}
}
\description{
# TODO FIND CLOSEST COMPANY IF group_cols=NULL
}
