% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preproc_analysis.R
\name{create_analysis_data}
\alias{create_analysis_data}
\title{Title}
\usage{
create_analysis_data(portfolio_data, multi_crispy, portfolio_crispy_merge_cols)
}
\arguments{
\item{multi_crispy}{}
}
\description{
\subsection{TODO HOW TO HANDLE ENTRIES IN PORTFOLIO, THAT ARE MISSING IN CRISPY ?}{
}
}
