% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_analysis.R
\name{load_input_plots_data}
\alias{load_input_plots_data}
\title{Data load function to generate plots}
\usage{
load_input_plots_data(
  trisk_output_dir,
  portfolio_data_path = NULL,
  granularity = c("company_id", "company_name", "ald_sector", "ald_business_unit"),
  maturity_month_term_bridge_fp = here::here("data", "maturity_month_term_bridge.csv"),
  trisk_start_year = 2022
)
}
\arguments{
\item{maturity_month_term_bridge_fp}{}
}
\description{
The dataframe in output of this function should always be
the one used as input for the plots preprocessing functions
}
