% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_main.R
\name{load_input_plots_data_from_files}
\alias{load_input_plots_data_from_files}
\title{Data load function to generate plots}
\usage{
load_input_plots_data_from_files(
  crispy_outputs_dir,
  portfolio_data_path = NULL,
  granularity = c("company_id", "company_name", "ald_sector", "ald_business_unit"),
  filter_outliers = FALSE
)
}
\arguments{
\item{crispy_outputs_dir}{crispy_outputs_dir}

\item{portfolio_data_path}{portfolio_data_path}

\item{granularity}{granularity}

\item{filter_outliers}{filter_outliers}
}
\description{
The dataframe in output of this function should always be
the one used as input for the plots preprocessing functions
}
