% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_data_load.R
\name{load_input_plots_data_from_tibble}
\alias{load_input_plots_data_from_tibble}
\title{Data load function to generate plots}
\usage{
load_input_plots_data_from_tibble(
  multi_crispy_data,
  portfolio_data = tibble::tibble(),
  granularity = c("company_id", "company_name", "ald_sector", "ald_business_unit"),
  trisk_start_year = NA
)
}
\arguments{
\item{granularity}{granularity}

\item{trisk_start_year}{(default) sets to the earliest year of multi_cripy_data}

\item{crispy_outputs_dir}{crispy_outputs_dir}

\item{portfolio_data_path}{portfolio_data_path}
}
\description{
The dataframe in output of this function should always be
the one used as input for the plots preprocessing functions
}
