% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_main.R
\name{main_data_load_trajectories_data}
\alias{main_data_load_trajectories_data}
\title{Title}
\usage{
main_data_load_trajectories_data(
  company_trajectories_data,
  granularity,
  param_cols = c("run_id", "scenario_geography", "baseline_scenario", "shock_scenario",
    "year")
)
}
\arguments{
\item{company_trajectories_data}{company_trajectories_data}

\item{max_trajectories_granularity}{max_trajectories_granularity}
}
\description{
Title
}
