% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_main.R
\name{main_load_analysis_data}
\alias{main_load_analysis_data}
\title{Title}
\usage{
main_load_analysis_data(
  multi_crispy_data,
  portfolio_data,
  portfolio_crispy_merge_cols
)
}
\arguments{
\item{portfolio_data}{portfolio_data}

\item{portfolio_crispy_merge_cols}{portfolio_crispy_merge_cols}

\item{multi_crispy}{multi_crispy}
}
\description{
Title
}
