% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_data_load.R
\name{main_load_analysis_data}
\alias{main_load_analysis_data}
\title{Title}
\usage{
main_load_analysis_data(
  portfolio_data,
  multi_crispy_data,
  trisk_run_params,
  portfolio_crispy_merge_cols
)
}
\arguments{
\item{portfolio_crispy_merge_cols}{}
}
\description{
Title
}
