% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_analysis.R
\name{main_load_multi_crispy_data}
\alias{main_load_multi_crispy_data}
\title{Title}
\usage{
main_load_multi_crispy_data(trisk_output_dir, max_crispy_granularity)
}
\arguments{
\item{max_crispy_granularity}{}
}
\description{
Title
}
