% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_main.R
\name{main_load_multi_crispy_data}
\alias{main_load_multi_crispy_data}
\title{Title}
\usage{
main_load_multi_crispy_data(multi_crispy_data, max_crispy_granularity)
}
\arguments{
\item{max_crispy_granularity}{max_crispy_granularity}

\item{crispy_outputs_dir}{crispy_outputs_dir}
}
\description{
Title
}
