% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_main.R
\name{main_load_multi_crispy_data}
\alias{main_load_multi_crispy_data}
\title{Title}
\usage{
main_load_multi_crispy_data(
  multi_crispy_data,
  granularity,
  param_cols = c("term", "run_id", "roll_up_type", "scenario_geography",
    "baseline_scenario", "shock_scenario", "risk_free_rate", "discount_rate",
    "div_netprofit_prop_coef", "carbon_price_model", "market_passthrough",
    "financial_stimulus", "start_year", "growth_rate", "shock_year")
)
}
\arguments{
\item{param_cols}{}
}
\description{
Title
}
