% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_main.R
\name{main_load_portfolio_data}
\alias{main_load_portfolio_data}
\title{Title}
\usage{
main_load_portfolio_data(
  portfolio_data,
  granularity,
  param_cols = c("portfolio_id", "term"),
  trisk_start_year
)
}
\arguments{
\item{trisk_start_year}{}
}
\description{
Title
}
