% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_main.R
\name{main_load_portfolio_data}
\alias{main_load_portfolio_data}
\title{Title}
\usage{
main_load_portfolio_data(
  portfolio_data,
  max_portfolio_granularity,
  trisk_start_year
)
}
\arguments{
\item{max_portfolio_granularity}{max_portfolio_granularity}
}
\description{
Title
}
