% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cdi_plots.R
\name{make_expected_loss_plot}
\alias{make_expected_loss_plot}
\title{Title}
\usage{
make_expected_loss_plot(
  data_cdi_el_plot,
  x_var,
  y_var,
  fill_var,
  facet_var,
  title
)
}
\arguments{
\item{data_cdi_el_plot}{data_cdi_el_plot}

\item{x_var}{x_var}

\item{y_var}{y_var}

\item{fill_var}{fill_var}

\item{title}{title}

\item{facet_rows_var}{facet_rows_var}

\item{facet_cols_var}{facet_cols_var}
}
\description{
Title
}
