% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_portfolio.R
\name{map_portfolio_maturity_to_term}
\alias{map_portfolio_maturity_to_term}
\title{Convert column year to term}
\usage{
map_portfolio_maturity_to_term(
  portfolio_data,
  maturity_month_term_bridge,
  trisk_start_year
)
}
\arguments{
\item{portfolio_data}{portfolio_data}

\item{maturity_month_term_bridge}{maturity_month_term_bridge}

\item{start_year}{start_year}
}
\description{
Convert column year to term
}
