% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_worst_performers.R
\name{qplot_worst_performers}
\alias{qplot_worst_performers}
\title{Title}
\usage{
qplot_worst_performers(data_performers_plot, x_var, y_var, fill_var, facet_var)
}
\arguments{
\item{facet_var}{}
}
\description{
Title
}
