% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_crispy.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Title}
\usage{
remove_outliers(
  df,
  column_filtered,
  index_columns,
  max_zscore = 3,
  min_obs = 30
)
}
\arguments{
\item{df}{df}

\item{column_filtered}{column name}

\item{index_columns}{index_columns}

\item{max_zscore}{max_zscore}

\item{min_obs}{min_obs}
}
\description{
Function to remove outliers based on z-score
}
